+++ GitGoingApp/.expo/packager-info.json
@@ -1,10 +1,9 @@
 {
   "devToolsPort": 19002,
-  "expoServerPort": 19000,
-  "packagerPort": 19001,
-  "packagerPid": 13884,
-  "expoServerNgrokUrl": "https://7m-qeq.anonymous.gitgoingapp.exp.direct",
-  "packagerNgrokUrl": "https://packager.7m-qeq.anonymous.gitgoingapp.exp.direct",
-  "ngrokPid": 12836
-
+  "expoServerPort": null,
+  "packagerPort": null,
+  "packagerPid": null,
+  "expoServerNgrokUrl": null,
+  "packagerNgrokUrl": null,
+  "ngrokPid": null
 }
