+++ GitGoingApp/.expo/settings.json
@@ -1,5 +1,5 @@
 {
-  "hostType": "localhost",
+  "hostType": "lan",
   "lanType": "ip",
   "dev": true,
   "minify": false,
