+++ GitGoingApp/src/GitClone.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { TouchableOpacity, Text, Button, StyleSheet, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import AntiClippy from './AntiClippy';
-
+//route to status
 const GitClone = () => {
-   const goToPull = () => {
-      Actions.gitPull()
+   const goToStatus = () => {
+      Actions.gitStatus()
    }
    return (
       <View style={styles.container}>
@@ -19,8 +19,8 @@ const GitClone = () => {
       </View>
       <TouchableOpacity style={styles.button}>
          <Button 
-         onPress = {goToPull}
-         title= "GitGoing to Git Pull"
+         onPress = {goToStatus}
+         title= "GitGoing to Git Status"
          color='#FF6D70'
          />
       </TouchableOpacity>
