+++ GitGoingApp/src/GitCommit.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { TouchableOpacity, Text, Button, StyleSheet, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import AntiClippy from './AntiClippy';
-
+//route ro reset
 const GitCommit = () => {
-   const goToStatus = () => {
-      Actions.gitStatus()
+   const goToReset = () => {
+      Actions.gitReset()
    }
    return (
       <View style={styles.container}>
@@ -19,8 +19,8 @@ const GitCommit = () => {
       </View>
       <TouchableOpacity style={styles.button}>
          <Button 
-         onPress = {goToStatus}
-         title= "GitGoing to Git Status"
+         onPress = {goToReset}
+         title= "GitGoing to Git Reset"
          color='#FF6D70'
          />
       </TouchableOpacity>
