+++ GitGoingApp/src/GitPull.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { TouchableOpacity, Text, Button, StyleSheet, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import AntiClippy from './AntiClippy';
-
+//route to add
 const GitPull = () => {
-   const goToHome = () => {
-      Actions.home()
+   const goToAdd = () => {
+      Actions.gitAdd()
    }
    return (
       <View style={styles.container}>
@@ -19,8 +19,8 @@ const GitPull = () => {
       </View>
       <TouchableOpacity style={styles.button}>
          <Button 
-         onPress = {goToHome}
-         title= "GitGoing Home"
+         onPress = {goToAdd}
+         title= "GitGoing to Git Add"
          color='#FF6D70'
          />
       </TouchableOpacity>
