+++ GitGoingApp/src/GitReset.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { TouchableOpacity, Text, Button, StyleSheet, View } from 'react-native'
 import { Actions } from 'react-native-router-flux'
 import AntiClippy from './AntiClippy'
-
+//route to push
 const GitReset = () => {
-   const goToStash = () => {
-      Actions.gitStash()
+   const goToPush = () => {
+      Actions.gitPush()
    }
    return (
       <View style={styles.container}>
@@ -19,8 +19,8 @@ const GitReset = () => {
       </View>
       <TouchableOpacity style={styles.button}>
          <Button 
-         onPress = {goToStash}
-         title= "GitGoing to Git Stash"
+         onPress = {goToPush}
+         title= "GitGoing to Git Push"
          color='#FF6D70'
          />
       </TouchableOpacity>
